backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
opencv-python-headless